fix component load
